Fix lint errors

+++ onlinejudges/codeforces/theatre_square/golang/main.go
@@ -10,8 +10,8 @@ func main() {
 	fmt.Scanf("%f%f%f", &n, &m, &a)
 
 	var w, h float64
-	w = math.Ceil(float64(n) / float64(a))
-	h = math.Ceil(float64(m) / float64(a))
+	w = math.Ceil(n / a)
+	h = math.Ceil(m / a)
 	r = w * h
 	fmt.Printf("%.0f\n", r)
 }
